Adding Dockerfile
Adding a debian based Dockerfile to build Pootle 2.7.0 docker images from

Using docker would:

 - Allow us to create/test standardised distro
 - Allow easy testing of multiple versions and/or migrations
 - Ease installation for new/prospective users

This Dockerfile uses sqlite so is not production-focused - we could fairly
easily add mysql/postgres versions

I wasn't able to get refresh_stats to run synchronously so it has to be run at
first run. This is not ideal and we could run this step by hand if we were to
make a distro image from it.

I have not added any VOLUMES atm - but we would probably want to.

Some documentation would probably be useful too.